Time: 0 ms (100.00%) | Memory: 23 MB (58.64%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int minEatingSpeed(vector<int>& piles, int h) {
-             int start=1;
-             int end=*max_element(piles.begin(), piles.end());
+             long long start=1;
+             long long end=*max_element(piles.begin(), piles.end());
           while(start<=end){
-            int mid=(start+end)/2;
+            long long mid=(start+end)/2;
         long long hour=0;
 
-        for(int pile : piles){
+        for(long long  pile : piles){
             hour+=(long long)(pile+mid-1)/mid;
         }
         if(hour<=h){
